[front-end] live update on comments
