+++ client/components/CommentForm.js
@@ -1,7 +1,5 @@
 import axios from 'axios';
-import homepage from '../pages/homepage'
-import auth from '../utils/auth';
-import sessions from '../utils/sessions';
+import homepage from '../pages/homepage';
 
 class CommentForm extends React.Component {
   constructor(props) {
@@ -18,27 +16,11 @@ class CommentForm extends React.Component {
   }
 
   postCommentWithText = () => {
-    this.postComment(this.state.commentText)
+    this.props.postComment(this.state.commentText)
   }
 
-  postComment = (commentText) => {
-    const token = sessions.getToken('jwt');
-    axios.get('/api', auth.makeTokenHeader(token))
-      .then(res => {
-        res.data.forEach(data => {
-          if (this.props.postTitle === data.title) {
-            return this.setState({ postId: data._id })
-          }
-        })
-        return this.state.postId
-      })
-      .then((res) => {
-        return axios.post(`/api/post/${this.state.postId}`, { body: commentText }, auth.makeTokenHeader(token))
-      })
-      .then((res) => {
-        console.log('SUCCESSFUL COMMENT POST')
-      })
-  }
+
+
 
   render() {
     return (
