+++ client/components/ParentPost.js
@@ -1,7 +1,3 @@
-import Link from 'next/link';
-import SubRedditBar from './SubredditBar';
-import Rating from './Rating';
-import React from 'react';
 import axios from 'axios';
 import ReactMarkdown from 'react-markdown';
 import Paper from 'material-ui/Paper';
@@ -9,7 +5,11 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import Postdetail from '../pages/postDetail';
 import CommentForm from './CommentForm';
 import CommentList from './CommentList';
-
+import auth from '../utils/auth';
+import sessions from '../utils/sessions';
+import Link from 'next/link';
+import SubRedditBar from './SubredditBar';
+import Rating from './Rating';
 
 class ParentPost extends React.Component {
   constructor(props) {
@@ -17,28 +17,48 @@ class ParentPost extends React.Component {
     this.state = {
       subredditId: '',
       postId: '',
+      commentBody: '',
       postTitle: '',
       postBodyText: '',
-      commentBody: '',
       comments: [],
     }
-
   }
 
   componentDidMount() {
-    // axios.get('/api')
-    //   .then(res => {
-    //     console.log('RESSS', res)
-    //   })
     axios.get('/api/post/5a8e0e2b7f911450d4600d99')
       .then(res => {
-        // console.log('res from parentpost', res)
         this.setState({
           postTitle: res.data.title,
           postBodyText: res.data.body,
           postId: res.data._id,
           comments: res.data.comments,
         })
+      })
+  }
+
+  postComment = (commentText) => {
+    const token = sessions.getToken('jwt');
+    axios.get('/api', auth.makeTokenHeader(token))
+      .then(res => {
+        res.data.forEach(data => {
+          if (this.state.postTitle === data.title) {
+            console.log('data', data)
+            return this.setState({ postId: data._id })
+          }
+        })
+        return this.state.postId
+      })
+      .then((res) => {
+        return axios.post(`/api/post/${this.state.postId}`, { body: commentText }, auth.makeTokenHeader(token))
+      })
+      .then((res) => {
+        console.log('SUCCESSFUL COMMENT POST')
+        return res;
+      })
+      .then(res => {
+        res.data.comments = [];
+        let newCommentArr = this.state.comments.push(res.data)
+        this.setState({ commentBody: newCommentArr })
       })
   }
 
@@ -54,16 +74,19 @@ class ParentPost extends React.Component {
               <ReactMarkdown source={this.state.postBodyText} />
             </Paper> <br />
           </div>
-        </MuiThemeProvider>
-        Add a new comment
-        <CommentForm title={this.state.title} subredditId={this.state.subredditId} postTitle={this.state.postTitle} />
-        {this.state.comments.body}
-        <CommentList comments={this.state.comments} />
-        <style>{`
+
+
+          Add a new comment
+        <CommentForm title={this.state.title} subredditId={this.state.subredditId} postComment={this.postComment} />
+
+          <CommentList comments={this.state.comments} newCommentBody={this.state.commentBody} />
+
+          <style>{`
           .postBody {
             padding: 4px 0 20px 0;
           }
         `}</style>
+        </MuiThemeProvider>
       </div>
